ALAudio: Use linear distance model.
It looks like this is the way UT expects things to work. Radius is
the absolute cutoff point (unlike UE2+), and a linear model is the
only model that provides this.

Not working as expected, probably because of the same underlying
problem I tried to solve with the strange AL_MAX_GAIN workaround.